stem tests: unify rose_stem teardown logic
* Move `rose_stem` test fixture into the conftest.py with the others
* Unify with the `test_dir` fixture (for provisioning the Cylc run dir)
  that all the other fixtures use.
* Remove the resource teardown logic from the `rose_stem` fixture,
  fixing a bug where the source directories from successful tests could
  be left behind.
